feat: create website with Mantine (#2)
* Setup Mantine

Follow the Mantine "Getting started" steps.

Also, I had to add the followin to `next.config.mjs`:

```
experimental: {
    optimizePackageImports: ["@mantine/core", "@mantine/hooks"],
},
```

It needed to be added to fix this warning: `<w>
[webpack.cache.PackFileCacheStrategy] Serializing big strings (111kiB)
impacts deserialization performance (consider using Buffer instead and
decode when needed)`.

* Add Vercel analytics

* Add BaseAutocomplete that can be extended

`BaseAutocomplete` can be used for autocompletes for the different data
needed for a course. The plan is to have an autocomplete for the
department, course number, and course section.

* Add onMatch and onBlur for Autocomplete

These callbacks are used to request data from the API as soon as
possible.

* Connect department and course number autocomplete

The course number autocomplete state now depends on the department
autocomplete.

* CourseAutocomplete currently not updating at times

If `departmentSelected` doesn't change from `true` to `false` and then
back to `true`, `CourseAutocomplete` won't update its disabled state.
This is a bug that needs to be fixed.

* Fix disabled state of course autocomplete

Now the disabled state is correctly set based on the department
autocomplete.

* Refactor to use onChange for autocomplete

Now the autocomplete feels more responsive.

* Add autocomplete loading text

Make it clear when the autocomplete is loading (making an API request).

* Reset course number error on department change

If the course number autocomplete has an error and the department
changes, the course number error correctly goes away. Refactoring is
also done to make the placeholder selection code cleaner.

* Make course numbers load on any department change

If the department was already selected, and the department was changed,
the course number autocomplete would not show that it is loading. This
issue is now fixed.

* Add section autocomplete

The section autocomplete depends on the department and course number
autocomplete.

* Work on getting Tailwind CSS to Work

Currently Tailwind CSS does not work.

* Fix tailwindcss not working

The issue was that `postcss.config.mjs` was missing `export default
config;`. It was accidentally removed in an earlier commit.

* Display selected course

The course the user has selected is now displayed.

* Add website header

The website now has a Mantine header component.

* Make course selector use row and default dark mode

The course selector now uses a row layout if there is space. Also, the
default colour theme is now dark mode.

* Add about page content

The about page now includes the following:
- Info
- Source code
- Getting course data (the npm package)

* Remove container component

The container component is not used anymore.

* Fix infinite loading bug

If the user types an entire autocomplete suggestion, and then still
clicks on the autocomplete suggestion, the next component loads forever.
For example, if the user types "CMPT" for department, then clicks on
"CMPT" from the autocomplete suggestion, then the course number
autocomplete would infinitely load.

* Add added courses list

The user can add a selected course to a list of added courses.

* Add year and term selectors

The user can now select the year and term to search for courses.

* Remove API error placeholder text

The API error message is only displayed below the entry box now. The
placeholder error text made it a bit confusing, since it looked like it
wasn't placeholder text because the colour was bright and not dull.

* Add Next.js website files and combobox example

Next.js website files include manifest data added, and various default
pages such as for:
- Not found (error 404)
- Manifest
- Loading

Also, the Mantine combobox example has been added. It will be used for
year and term selection.

* Implement year and term dropdowns

The year and term dropdowns now connect with the API to select a year
and term.

* Fix course selection state bugs

The text and states (loading, disabled) wouldn't always be set
correctly. Update how some states are set.

* Download example calendar file

The download calendar file now downloads an example calendar file. It
doesn't use the added courses yet.

* Prepare for implementing calendar file download

Investigate ical-generator.

* Add year and term error style and label

The combobox for the year and term selection now matches the error and
label style of the built-in Mantine autocomplete.

* Clean up term selector combobox style

Don't use `style` and instead use `styles` for style.

* Update README and about page

The README now uses a nice fancy template.

* Add opengraph image

Add the image shown during social shares.

* Fix incorrect urls

Some urls that were incorrectly set are now fixed.

* Add file import tutorial section

This new section will be updated with instructions for how to import the
downloaded calendar file into a calendar.

* Potential fix for no ordered tutorial list

Can turn off Tailwind CSS preflight. However, this messes up the Spotify
artist embed.

* Keep Tailwind CSS preflight and fix tutorial list

The tutorial list ordered style is now done with Tailwind CSS.

* Update import tutorial steps

Make the steps clearer.